Subject: Added SetSelectedResource method and updated related classes
Extended the `EditorManager` class with a new method `SetSelectedResource` to set the selected resource in the `DataEditorWindow`. Updated `DataEditorWindow` class to include `EditorManager` and `SceneHierarchyWindow` headers and modified `SetSelectedEntity` method. Added a new pure virtual method `SetSelectedResource` in the `Application` interface, requiring all implementing classes to provide an implementation. Updated `ResourceManager` class to call `SetSelectedResource` on the `Editor` instance when a resource is selected, allowing the selected resource to be updated in the `Editor`.