Implemented entity copying functionality
In this commit, we have introduced the ability to copy entities. The `CreateEntityCopy(Entity* entity)` method has been added to the `EditorManager`, `Application`, and `SceneHierarchyWindow` classes, and declared as virtual in the latter two. A "Paste Entity" button has been added to the `SceneHierarchyWindow` interface, along with methods to handle shortcut inputs and create entity copies. The `CopyEntityCommand` class has been implemented to handle the execution, undoing, and redoing of entity copy commands. The `CopyComponent(Component* comp, std::string entityID)` method has been added to the `Component`, `Datalist`, and `ComponentDatalist` classes to facilitate component copying. Changes have been made to various files to handle the copying of components, including the addition of copy assignment operators and modification of default field values. The `Pipeline` default values in `TechDemo1.npj` and `Pipeline.h` have been updated. Lastly, the entity dropdown and resource selection interfaces in `ToolUI.cpp` have been updated to handle the copying of entities and resources.