`Added HandleShortcutInputs method in SceneViewWindow`
` `

`A new method, HandleShortcutInputs(), was added to the SceneViewWindow class in SceneViewWindow.cpp. This method retrieves the EditorManager and SceneHierarchyWindow instances, and then calls the HandleShortcutInputs() method on the SceneHierarchyWindow instance. The declaration of this new method was also added to the public section of the SceneViewWindow class in SceneViewWindow.h.