Updated entity copying
In `SceneHierarchyWindow.cpp` and `CommandTypes.cpp`, the `CopyEntityCommand` now takes a vector of `EntityCopy` objects, allowing for multiple entities to be copied at once. The `CopyEntityCommand::Execute` method processes each `EntityCopy` in the vector and stores the resulting copied entities.

The `EntityCopy` struct in `CommandTypes.h` has been updated to include a vector of `EntityCopy` objects representing the children of the entity being copied.

Finally, an assignment operator has been added to the `Transform` struct in `Transform.h`, allowing one `Transform` object to be assigned the values of another.