Updated entity copy conditions, and component addition logic
In `CommandTypes.cpp`, added a condition in `ProcessEntityCopy` method to check if `m_sEntityParentID` of `copiedEntity` is empty. In `ComponentList.hpp`, updated `AddComponent` method to store `m_sEntityID` of `comp` object in a local variable and set `m_bAvailableForReuse` of `co` object to `false`. Removed the return statement after `m_deletedComponents.pop_front()`, allowing the method to continue execution.