Updated Profiler and PerformanceStats for enhanced memory tracking
Significant changes were made to the Profiler and PerformanceStats classes to improve memory usage tracking. In `Profiler.cpp`, the initialization of `m_pStats` and `m_vFrameTimeStats` was moved to a new function `Profiler::ProcessData`. The `Profiler::DoInterface` function was heavily modified to move calculations to `Profiler::ProcessData` and add a new section for memory usage stats. A new struct `MemoryStat` was added to `Profiler.h`. The `Application::ClearLoadedScene` function was updated to clear the `m_onScreenObjects` vector, fixing a crash that could occur when changing scenes with the profiler open. The `PerformanceStats::UpdateMemoryUsageStats` function now calculates memory usage stats in MB and available physical and virtual memory. New functions `GetPhysicalMemoryAvailable` and `GetVirtualMemoryAvailable` were added to `PerformanceStats.cpp`. The `PerformanceStats` class was updated to change the data type of memory usage stats to `float` and add new members for available memory.